v3

+++ RealmOfTheGuardians/src/tk/damnesia/entity/Animation.java
@@ -1,13 +1,7 @@
-// Decompiled by Jad v1.5.8g. Copyright 2001 Pavel Kouznetsov.
-// Jad home page: http://www.kpdus.com/jad.html
-// Decompiler options: packimports(3) 
-// Source File Name:   Animation.java
-
 package tk.damnesia.entity;
 
 import java.awt.Graphics;
 import java.awt.Image;
-import java.io.PrintStream;
 import java.util.ArrayList;
 
 public class Animation {
@@ -32,7 +26,7 @@ public class Animation {
 	}
 
 	public void update() {
-		if (System.currentTimeMillis() - lastLoopTime.longValue() > (long) looptime) {
+		if (System.currentTimeMillis() - lastLoopTime.longValue() > looptime) {
 			lastLoopTime = Long.valueOf(System.currentTimeMillis());
 			if (index < images.size() - 1)
 				index++;
